% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betti_criteria.R
\name{bettiCriteria}
\alias{bettiCriteria}
\title{For a specific topological subspace in the filtration, this function calculate the betti differences between persistence diagrams D1 and D2 produced by two datasets (which are typically considered to be an original dataset S and its perturbation S').}
\usage{
bettiCriteria(D1_dim, D2_dim, point)
}
\description{
param D1_dim		A matrix with persistence diagrams for a dataset X
param D2_dim		A matrix with persistence diagrams for a dataset X'
param point		A specific point in the considered filtration
return		A absolute difference of betti numbers associated with the topological subspace X_{point}
}
