% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betti_criteria.R
\name{bettiPerturbation}
\alias{bettiPerturbation}
\title{This function calculates the betti differences between the persistence diagram created with an original dataset S and other created by reducing the size of S.}
\usage{
bettiPerturbation(dat, nLandmarks, maxDimension, isWeak, relaxed, alpha,
  maxscale, nFiltration, nRemoving)
}
\description{
param X		Input data points matrix
param nLandmarks	Number of landmarks adopted in lazy witness which are chosen by random subsampling
param maxDimension	Maximum dimesion on the constructed complex
param relaxed	If true the lazy witness is relaxed as shown in http://gudhi.gforge.inria.fr/doc/latest/group__witness__complex.html and alpha applies. 
param alpha		Parameter alpha of relaxed lazy witness
param maxscale	The maxscale parameter employed to produce the filtration of witness complexes (of GUDHI package)
param nFiltration	Number of topological subspace in the considered filtration
param nRemoving	Number of elements to be removed from the dataset for new calculation of the persistence diagram
return		A persistence diagram
}
