% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betti_criteria.R
\name{datasetSubsampling}
\alias{datasetSubsampling}
\title{Remove a number of data points in a dataset}
\usage{
datasetSubsampling(X, npoints)
}
\description{
param X		The chosen dataset
param npoints	The number of points to remove
return		The new dataset
}
