% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betti_criteria.R
\name{seriesBettiCriteria}
\alias{seriesBettiCriteria}
\title{This function calculates the betti differences between persistence diagrams D1 and D2 produced by two datasets (which are typically considered to be an original dataset S and its perturbation S'). Given a topological space $X$ and a tame function $f:\mathbb R \to X$ there is a filtration $X_0 \subseteq X_1 \subseteq \dots \subseteq X_t$. This function will calculated the absolute betti number differences between D1 and D2 for each topological subspace $X_i$ with $0 \leq i \leq t$.}
\usage{
seriesBettiCriteria(D1, D2, dim, maxscale, nFiltration)
}
\description{
param D1		A matrix with persistence diagrams for a dataset X
param D2		A matrix with persistence diagrams for a dataset X'
param dim		Dimension of the simplicial complexes
param maxscale	The maxscale parameter employed to produce the filtration of witness complexes (of GUDHI package)
param nFiltration	Number of topological subspace in the considered filtration
return		A absolute difference of betti numbers associated with each topological subspace in the filtration
}
